#57 Refactor the Spotify API service into separate services for auth, controls, and playback polling
