+++ src/app/components/track-player/track-player-progress/track-player-progress.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { MatSliderChange } from '@angular/material/slider';
-import { SpotifyService } from '../../../services/spotify/spotify.service';
+import { SpotifyControlsService } from '../../../services/spotify/controls/spotify-controls.service';
 
 @Component({
   selector: 'app-track-player-progress',
@@ -15,7 +15,7 @@ export class TrackPlayerProgressComponent implements OnChanges {
   progressFormatted = this.getFormattedProgress(this.progress);
   durationFormatted = this.getFormattedProgress(this.duration);
 
-  constructor(private spotify: SpotifyService) {}
+  constructor(private controls: SpotifyControlsService) {}
 
   ngOnChanges(changes: SimpleChanges): void {
     if (changes.progress) {
@@ -27,7 +27,7 @@ export class TrackPlayerProgressComponent implements OnChanges {
   }
 
   onProgressChange(change: MatSliderChange): void {
-    this.spotify.setTrackPosition(change.value);
+    this.controls.setTrackPosition(change.value);
   }
 
   private getFormattedProgress(milliseconds: number): string {
